requirements.txt
Babel==1.3
DateTime==4.0.1
Django==1.6.5
South==1.0
argparse==1.2.1
django-object-tools==0.0.7
django-pgfields==1.4.4
django-south-admin==0.0.3
httplib2==0.9
iso8601==0.1.10
magicmock==0.03
mock==1.0.1
netaddr==0.7.11
oath==1.2
oauthlib==0.6.3
oslo.config==1.3.0
pbr==0.8.2
prettytable==0.7.2
psycopg2==2.5.3
python-keystoneclient==0.9.0
python-swiftclient==2.1.0
pytz==2014.4
requests==2.3.0
requests-oauthlib==0.4.1
simplejson==3.5.3
six==1.7.2
uuid==1.30
zope.interface==4.1.1
easywebdav==1.2.0
pycrypto==2.6.1